Move dunit test to geode-core
