+++ geode-connectors/build.gradle
@@ -28,23 +28,6 @@ repositories {
 
 apply from: "${project.projectDir}/../gradle/publish.gradle"
 
-configurations {
-   //Configuration used to download mysql jar
-   jdbcTestingJars
-}
-def jdbcJarsDir = "$buildDir/generated-resources/jdbcJars"
-sourceSets {
-  distributedTest {
-    resources {
-      srcDirs jdbcJarsDir
-    }
-    output.dir(jdbcJarsDir, builtBy: 'downloadJdbcJars')
-  }
-}
-task downloadJdbcJars(type:Copy) {
-  from {configurations.jdbcTestingJars}
-  into jdbcJarsDir
-}
 
 dependencies {
   compile(platform(project(':boms:geode-all-bom')))
@@ -117,10 +100,6 @@ dependencies {
   acceptanceTestRuntime('mysql:mysql-connector-java')
   acceptanceTestRuntime('org.apache.derby:derby')
   acceptanceTestRuntime('org.postgresql:postgresql')
-
-  jdbcTestingJars('mysql:mysql-connector-java:8.0.15')
 }
 
 integrationTest.forkEvery 0
-
-
